requirements.txt
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
torch==2.2.1
torchmetrics==1.4.0.post0
torchsummary==1.5.1
torchvision==0.17.1
lightning==2.3.3
numpy==1.26.4
matplotlib==3.8.3
albumentations==1.4.11
tqdm==4.66.2
pyyaml==6.0.1